requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
azure-storage-file-datalake==12.14.0
azure-identity==1.15.0
pytest==7.4.3
pytest-mock==3.12.0
python-dotenv==1.0.0
